requirements.txt
fastapi["standard"]
python-dotenv
redis
sqlmodel